requirements.txt
python-telegram-bot==20.7
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
schedule==1.2.0
